requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
openai>=1.0.0
aiohttp>=3.9.0
lxml>=4.9.0
selenium>=4.15.0
webdriver-manager>=4.0.0
feedparser>=6.0.10
python-dateutil>=2.8.0
tqdm>=4.66.0
fake-useragent>=1.4.0
retry>=0.9.2
